Удаляет модуль store и возвращает логику инициализации в index
